django-on-docker/app/requirements.txt
Pillow==7.1.2
Django==3.0.6
psycopg2-binary==2.8.3
djangorestframework
